I've been working on making the Move/Copy function object oriented. This has proven to be a vastly more involved project than anticipated. During the course of this I have had several major problems that I am going to attempt to describe to you here. I have succeeded in getting stories and pages to move, but not yet sections. The function controlling all of this business is copyObj() in the segue class. It is being called from copyParts.php. Maybe I am just crazy (made so by beating my head off of my desk every day this week), but some of the changes I have made worry me and I was hoping for an opinion on them.
1. Fetching Down.
	I changed all of the fetchDown functions to force a fetch if $full is specified. I was having lots of problems with objects
being listed as fetcheddown, but not actually fetched down.  As I'm not sure of the function
of the fetcheddown flag, I was loath to mess with it, however, maybe it should only be set if the object is fetched in full.

2. InsertDB.
	Moving/Copying an object involves taking an object that exists, with children, then inserting it into a new location, then
inserting its children into new locations, then adding the children to the children array. What insertDB didn't do is remove the
references to the old children, putting multiple copies of stories on pages.  I accomplished this removal of the old child references
by adding a "database delete" option to the various deleteChild functions that is by default true, and only false when "cleaning up" of
the children when insertDBing. Hopefully this won't hurt anything, but again I feel like this is kind of fishy. Maybe I should write a
new function for each

3. Tables.
	This isn't really a problem, just a question.  as long as I have an empty permissions table in the database, I can replace all
of the other tables without harm right?  the data-integrity on etdev is all crapped up and I am worried that I might to try to fix an
error that is only present because the tables are screwed up from previous bugs.

+++ objects/page.inc.php
@@ -65,15 +65,17 @@ class page extends segue {
 		$this->changed[stories]=1;
 	}
 	
-	function delStory($id) {
+	function delStory($id,$delete=1) {
 		$d = array();
 		foreach ($this->getField("stories") as $s) {
 			if ($s != $id) $d[]=$s;
 		}
 		$this->data[stories] = $d;
 		$this->changed[stories]=1;
-		$story = new story($this->owning_site,$this->owning_section,$this->owning_page,$id);
-		$story->delete();
+		if ($delete) {
+			$story = new story($this->owning_site,$this->owning_section,$this->owning_page,$id);
+			$story->delete();
+		}
 	}
 	
 	function fetchUp() {
@@ -87,8 +89,8 @@ class page extends segue {
 	}
 	
 	function fetchDown($full=0) {
-		if (!$this->fetcheddown) {
-/* 			print "---->page fetchdown".$this->id."<BR>"; */
+		if (!$this->fetcheddown || $full) {
+/* 			print "---->page fetchdown".$this->id." full = $full<BR>"; */
 			if (!$this->tobefetched || $full) $this->fetchFromDB(0,$full);
 			foreach ($this->getField("stories") as $s) {
 				$this->stories[$s] = new story($this->owning_site,$this->owning_section,$this->id,$s);
@@ -146,8 +148,14 @@ class page extends segue {
 	}
 	
 	function insertDB($down=0,$newsite=null,$newsection=0,$keepaddedby=0) {
+		$origsite = $this->owning_site;
+		$origid = $this->id;
 		if ($newsite) $this->owning_site = $newsite;
-		if ($newsection) $this->owning_section = $newsection;
+		if ($newsection) {
+			$this->owning_section = $newsection;
+			$this->owningSectionObj = new section($newsite,$newsection);
+		}
+		
 		$a = $this->createSQLArray(1);
 		if (!$keepaddedby) {
 			$a[] = "addedby='$_SESSION[auser]'";
@@ -158,13 +166,14 @@ class page extends segue {
 		}
 
 		$query = "insert into pages set ".implode(",",$a);
-		print $query; //debug
+		print $query."<br>"; //debug
 		db_query($query);
 		
 		$this->id = mysql_insert_id();
 		
 		$this->fetchUp();
 		$this->owningSectionObj->addPage($this->id);
+		$this->owningSectionObj->delPage($origid,0);
 		$this->owningSectionObj->updateDB();
 		
 		// add new permissions entry.. force update
@@ -180,7 +189,7 @@ class page extends segue {
 		return true;
 	}
 	
-	function createSQLArray() {
+	function createSQLArray($all=0) {
 		$d = $this->data;
 		$a = array();
